+++ src/components/model/film/FilmReviewListWithTop/index.ts
@@ -1,0 +1,1 @@
+export { default as FilmReviewListWithTop } from './FilmReviewListWithTop'
