+++ src/components/model/film/filmReviewList/FilmReviewList.tsx
@@ -1,5 +1,4 @@
-import { Divider, Theme } from '@mui/material'
-import { makeStyles } from '@mui/styles'
+import { Divider } from '@mui/material'
 import { Box } from '@mui/system'
 import React, { FC } from 'react'
 import { FilmReview } from '../filmReveiw'
@@ -14,13 +13,11 @@ const FilmReviewList: FC<Props> = ({ reviewList, isMobileSize }) => {
   return (
     <Box>
       {reviewList.map((review) => (
-        <>
-          <Box key={review.id}>
-            <FilmReview {...review} {...{ isMobileSize }} />
-            <Divider />
-            <Box padding={1} />
-          </Box>
-        </>
+        <Box key={review.id}>
+          <FilmReview {...review} {...{ isMobileSize }} />
+          <Divider />
+          <Box padding={1} />
+        </Box>
       ))}
     </Box>
   )
