--- src/components/organisms/film/FilmReviewList/index.ts
@@ -1,1 +1,0 @@
-export { default as FilmReviewList } from './FilmReviewList'
